requirements.txt
fastapi==0.116.1
uvicorn==0.24.0
sqlalchemy==2.0.21
pydantic==2.4.0
pymysql
python-dotenv==1.0.0
email-validator==2.1.0
